добавлена бизнес логика

+++ internal/models/models.go
@@ -1,8 +1,8 @@
 package models
 
 type BankRate struct {
-	ID          int64  `db:"id"`
-	BankName    string `db:"bank_name"`
-	DepositName string `db:"deposit_name"`
-	Rate        int    `db:"rate"`
+	ID          int64   `db:"id"`
+	BankName    string  `db:"bank_name"`
+	DepositName string  `db:"deposit_name"`
+	Rate        float32 `db:"rate"`
 }
